test: ModifyAlarmController, ModifyAlarmAdapter
