--- src/api/src/main/kotlin/com/tenmm/tilserver/alaram/adapter/inbound/model/GetAlarmResponse.kt
@@ -1,8 +1,0 @@
-package com.tenmm.tilserver.alaram.adapter.inbound.model
-
-data class GetAlarmResponse(
-    val userIdentifier: String,
-    val enable: Boolean,
-    val iteration: String
-
-)
