--- src/api/src/main/kotlin/com/tenmm/tilserver/alaram/adapter/inbound/model/ModifyAlarmRequest.kt
@@ -1,6 +1,0 @@
-package com.tenmm.tilserver.alaram.adapter.inbound.model
-
-data class ModifyAlarmRequest(
-    val enable: Boolean,
-    val iteration: String,
-)
