--- src/api/src/main/kotlin/com/tenmm/tilserver/alaram/adapter/inbound/model/ModifyAlarmResponse.kt
@@ -1,5 +1,0 @@
-package com.tenmm.tilserver.alaram.adapter.inbound.model
-
-data class ModifyAlarmResponse(
-    val isSuccess: Boolean,
-)
