--- src/api/src/main/kotlin/com/tenmm/tilserver/alaram/adapter/outbound/persistence/converter/model-converters.kt
@@ -1,19 +1,0 @@
-import com.tenmm.tilserver.alaram.domain.Alarm
-import com.tenmm.tilserver.common.domain.Identifier
-import com.tenmm.tilserver.outbound.persistence.entity.AlarmEntity
-
-fun Alarm.toEntity(): AlarmEntity {
-    return AlarmEntity(
-        userIdentifier = this.userIdentifier.toString(),
-        enable = this.enable,
-        iteration = this.iteration
-    )
-}
-
-fun AlarmEntity.toModel(): Alarm {
-    return Alarm(
-        userIdentifier = Identifier(this.userIdentifier),
-        enable = this.enable,
-        iteration = this.iteration
-    )
-}
