--- src/api/src/main/kotlin/com/tenmm/tilserver/alaram/application/inbound/ModifyAlarmUseCase.kt
@@ -1,7 +1,0 @@
-package com.tenmm.tilserver.alaram.application.inbound
-
-import com.tenmm.tilserver.alaram.application.inbound.model.ModifyAlarmCommand
-
-interface ModifyAlarmUseCase {
-    fun modifyAlarm(alarm: ModifyAlarmCommand): Boolean
-}
