--- src/api/src/main/kotlin/com/tenmm/tilserver/alaram/application/inbound/model/ModifyAlarmCommand.kt
@@ -1,9 +1,0 @@
-package com.tenmm.tilserver.alaram.application.inbound.model
-
-import com.tenmm.tilserver.common.domain.Identifier
-
-data class ModifyAlarmCommand(
-    val userIdentifier: Identifier,
-    val enable: Boolean,
-    val iteration: String,
-)
