--- src/api/src/main/kotlin/com/tenmm/tilserver/alaram/application/outbound/GetAlarmPort.kt
@@ -1,8 +1,0 @@
-package com.tenmm.tilserver.alaram.application.outbound
-
-import com.tenmm.tilserver.alaram.domain.Alarm
-import com.tenmm.tilserver.common.domain.Identifier
-
-interface GetAlarmPort {
-    fun findByUserIdentifier(userIdentifier: Identifier): Alarm?
-}
