--- src/api/src/main/kotlin/com/tenmm/tilserver/alaram/application/outbound/ModifyAlarmPort.kt
@@ -1,7 +1,0 @@
-package com.tenmm.tilserver.alaram.application.outbound
-
-import com.tenmm.tilserver.alaram.application.inbound.model.ModifyAlarmCommand
-
-interface ModifyAlarmPort {
-    fun modifyByUserIdentifier(command: ModifyAlarmCommand): Boolean
-}
