+++ src/api/src/main/kotlin/com/tenmm/tilserver/alarm/adapter/inbound/rest/model/GetAlarmResponse.kt
@@ -1,0 +1,8 @@
+package com.tenmm.tilserver.alarm.adapter.inbound.rest.model
+
+data class GetAlarmResponse(
+    val userIdentifier: String,
+    val enable: Boolean,
+    val iteration: String
+
+)
