+++ src/api/src/main/kotlin/com/tenmm/tilserver/alarm/adapter/inbound/rest/model/ModifyAlarmRequest.kt
@@ -1,0 +1,6 @@
+package com.tenmm.tilserver.alarm.adapter.inbound.rest.model
+
+data class ModifyAlarmRequest(
+    val enable: Boolean,
+    val iteration: String,
+)
