+++ src/api/src/main/kotlin/com/tenmm/tilserver/alarm/adapter/inbound/rest/model/ModifyAlarmResponse.kt
@@ -1,0 +1,5 @@
+package com.tenmm.tilserver.alarm.adapter.inbound.rest.model
+
+data class ModifyAlarmResponse(
+    val isSuccess: Boolean,
+)
