+++ src/api/src/main/kotlin/com/tenmm/tilserver/alarm/application/outbound/ModifyAlarmPort.kt
@@ -1,0 +1,7 @@
+package com.tenmm.tilserver.alarm.application.outbound
+
+import com.tenmm.tilserver.alarm.application.inbound.model.ModifyAlarmCommand
+
+interface ModifyAlarmPort {
+    fun modifyByUserIdentifier(command: ModifyAlarmCommand): Boolean
+}
