+++ src/api/src/test/kotlin/com/tenmm/tilserver/alarm/adapter/inbound/rest/GetAlarmControllerTest.kt
@@ -1,0 +1,7 @@
+package com.tenmm.tilserver.alarm.adapter.inbound.rest
+
+import io.mockk.junit5.MockKExtension
+import org.junit.jupiter.api.extension.ExtendWith
+
+@ExtendWith(MockKExtension::class)
+class GetAlarmControllerTest
