+++ src/api/src/test/kotlin/com/tenmm/tilserver/alarm/adapter/inbound/rest/model/randomAlarmInboundAdapterModel.kt
@@ -1,0 +1,8 @@
+package com.tenmm.tilserver.alarm.adapter.inbound.rest.model
+
+fun randomModifyAlarmRequest(isEnable: Boolean = true): ModifyAlarmRequest {
+    return ModifyAlarmRequest(
+        enable = isEnable,
+        iteration = "* * * * * *"
+    )
+}
